move client directory
