+++ src/components/auth/ResetPasswordForm.tsx
@@ -13,7 +13,7 @@ import Typography from "@mui/material/Typography";
 import { Controller, useForm } from "react-hook-form";
 import { z as zod } from "zod";
 
-import { authClient } from "@/utils/auth/client";
+import { authClient } from "@/utils/client";
 
 const schema = zod.object({
   email: zod.string().min(1, { message: "Email is required" }).email(),
