+++ src/components/auth/SignInForm.tsx
@@ -21,7 +21,7 @@ import { Eye as EyeIcon } from "@phosphor-icons/react/dist/ssr/Eye";
 import { EyeSlash as EyeSlashIcon } from "@phosphor-icons/react/dist/ssr/EyeSlash";
 
 import { paths } from "@/paths";
-import { authClient } from "@/utils/auth/client";
+import { authClient } from "@/utils/client";
 import { useUser } from "@/hooks/useUser";
 
 const schema = zod.object({
